# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Project OMNI Fitness OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workouts:
    get:
      operationId: listWorkouts
      tags: [workouts]
      summary: List all workouts
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of workouts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workout"
    post:
      operationId: createWorkout
      tags: [workouts]
      summary: Create a workout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkoutBody"
      responses:
        "201":
          description: Created workout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
  /workouts/{id}:
    get:
      operationId: getWorkout
      tags: [workouts]
      summary: Get a workout
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workout details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
    patch:
      operationId: updateWorkout
      tags: [workouts]
      summary: Update a workout
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkoutBody"
      responses:
        "200":
          description: Updated workout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
    delete:
      operationId: deleteWorkout
      tags: [workouts]
      summary: Delete a workout
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Workout:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        notes:
          type: string
          nullable: true
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        duration:
          type: integer
          description: Duration in seconds
          nullable: true
        sets:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutSet"
        totalVolume:
          type: number
          nullable: true
        totalSets:
          type: integer
          nullable: true
        routineId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - startedAt
        - createdAt
    WorkoutSet:
      type: object
      properties:
        id:
          type: integer
        exerciseId:
          type: integer
        exerciseName:
          type: string
        setNumber:
          type: integer
        setType:
          type: string
          enum: [warmup, normal, dropset, failure, negative]
        weight:
          type: number
          nullable: true
        reps:
          type: integer
          nullable: true
        duration:
          type: integer
          nullable: true
        distance:
          type: number
          nullable: true
        rpe:
          type: number
          minimum: 1
          maximum: 10
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - id
        - exerciseId
        - exerciseName
        - setNumber
        - setType
    CreateWorkoutBody:
      type: object
      properties:
        name:
          type: string
        notes:
          type: string
          nullable: true
        routineId:
          type: integer
          nullable: true
        sets:
          type: array
          items:
            $ref: "#/components/schemas/CreateSetBody"
      required:
        - name
    UpdateWorkoutBody:
      type: object
      properties:
        name:
          type: string
        notes:
          type: string
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        duration:
          type: integer
          nullable: true
        sets:
          type: array
          items:
            $ref: "#/components/schemas/CreateSetBody"
    CreateSetBody:
      type: object
      properties:
        exerciseId:
          type: integer
        setNumber:
          type: integer
        setType:
          type: string
          enum: [warmup, normal, dropset, failure, negative]
          default: normal
        weight:
          type: number
          nullable: true
        reps:
          type: integer
          nullable: true
        duration:
          type: integer
          nullable: true
        distance:
          type: number
          nullable: true
        rpe:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - exerciseId
        - setNumber
...